#216 + #217: respect React's Rules of Hooks
